+++ src-tauri/src/invokers/config_invokers.rs
@@ -1,8 +1,7 @@
-use quo_common::config::{DefaultValue, Setting, SETTINGS};
+use quo_common::config::{DefaultValue, Setting, AVAILABLE_THEMES, SETTINGS};
 use quo_common::QUO_CONFIG_STORE_NAME;
 use serde::Serialize;
 use serde_json::Value;
-use std::fs;
 use tauri_plugin_store::StoreExt;
 
 #[derive(Serialize)]
@@ -13,52 +12,8 @@ pub struct SettingDto {
 }
 
 #[tauri::command]
-pub fn get_available_themes(app: tauri::AppHandle) -> Result<Vec<String>, String> {
-    use tauri::Manager;
-    let mut themes = Vec::new();
-
-    // Try multiple possible paths to find src/styles
-    let mut paths = vec![
-        "../src/styles".to_string(),
-        "src/styles".to_string(),
-        "./src/styles".to_string(),
-    ];
-
-    if let Ok(resource_dir) = app.path().resource_dir() {
-        paths.push(resource_dir.join("src").join("styles").to_string_lossy().to_string());
-        paths.push(resource_dir.join("styles").to_string_lossy().to_string());
-    }
-
-    for path in paths {
-        if let Ok(entries) = fs::read_dir(path) {
-            for entry in entries.flatten() {
-                if let Ok(file_type) = entry.file_type() {
-                    if file_type.is_dir() {
-                        if let Some(name) = entry.file_name().to_str() {
-                            themes.push(name.to_string());
-                        }
-                    }
-                }
-            }
-            if !themes.is_empty() {
-                break;
-            }
-        }
-    }
-
-    if themes.is_empty() {
-        // Fallback to default themes if directory not found
-        themes.push("Quo (default)".to_string());
-    }
-
-    // Ensure they are sorted and "Quo (default)" is first if present
-    themes.sort();
-    if let Some(pos) = themes.iter().position(|t| t == "Quo (default)") {
-        let quo_default = themes.remove(pos);
-        themes.insert(0, quo_default);
-    }
-
-    Ok(themes)
+pub fn get_available_themes(_app: tauri::AppHandle) -> Result<Vec<String>, String> {
+    Ok(AVAILABLE_THEMES.iter().map(|&t| t.to_string()).collect())
 }
 
 #[tauri::command]
